refactor: move some pydantic models and metafiles
This should reduce some of the circular imports.
