+++ charmcraft/commands/pack.py
@@ -31,7 +31,7 @@ from charmcraft import env, parts, instrum
 from charmcraft.cmdbase import BaseCommand
 from charmcraft.commands import build
 from charmcraft.errors import DuplicateCharmsError
-from charmcraft.manifest import create_manifest
+from charmcraft.metafiles.manifest import create_manifest
 from charmcraft.parts import Step
 from charmcraft.utils import (
     load_yaml,
