+++ charmcraft/format.py
@@ -1,0 +1,105 @@
+# Copyright 2023 Canonical Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# For further info, check https://github.com/canonical/charmcraft
+
+"""Reformat pydantic errors."""
+
+from typing import Tuple
+
+
+def format_pydantic_error_location(loc) -> str:
+    """Format location."""
+    loc_parts = []
+    for loc_part in loc:
+        if isinstance(loc_part, str):
+            loc_parts.append(loc_part)
+        elif isinstance(loc_part, int):
+            # Integer indicates an index. Go
+            # back and fix up previous part.
+            previous_part = loc_parts.pop()
+            previous_part += f"[{loc_part}]"
+            loc_parts.append(previous_part)
+        else:
+            raise RuntimeError(f"unhandled loc: {loc_part}")
+
+    loc = ".".join(loc_parts)
+
+    # Filter out internal __root__ detail.
+    loc = loc.replace(".__root__", "")
+    return loc
+
+
+def format_pydantic_error_message(msg: str) -> str:
+    """Format pydantic's error message field."""
+    # Replace shorthand "str" with "string".
+    msg = msg.replace("str type expected", "string type expected")
+    return msg
+
+
+def printable_field_location_split(location: str) -> Tuple[str, str]:
+    """Return split field location.
+
+    If top-level, location is returned as unquoted "top-level".
+    If not top-level, location is returned as quoted location.
+
+    Examples:
+    (1) field1[idx].foo => 'foo', 'field1[idx]'
+    (2) field2 => 'field2', top-level
+
+    :returns: Tuple of <field name>, <location> as printable representations.
+    """
+    loc_split = location.split(".")
+    field_name = repr(loc_split.pop())
+
+    if loc_split:
+        return field_name, repr(".".join(loc_split))
+
+    return field_name, "top-level"
+
+
+def format_pydantic_errors(errors, *, file_name: str = "charmcraft.yaml") -> str:
+    """Format errors.
+
+    Example 1: Single error.
+
+    Bad charmcraft.yaml content:
+    - field: <some field>
+      reason: <some reason>
+
+    Example 2: Multiple errors.
+
+    Bad charmcraft.yaml content:
+    - field: <some field>
+      reason: <some reason>
+    - field: <some field 2>
+      reason: <some reason 2>
+    """
+    combined = [f"Bad {file_name} content:"]
+    for error in errors:
+        formatted_loc = format_pydantic_error_location(error["loc"])
+        formatted_msg = format_pydantic_error_message(error["msg"])
+
+        if formatted_msg == "field required":
+            field_name, location = printable_field_location_split(formatted_loc)
+            combined.append(f"- field {field_name} required in {location} configuration")
+        elif formatted_msg == "extra fields not permitted":
+            field_name, location = printable_field_location_split(formatted_loc)
+            combined.append(
+                f"- extra field {field_name} not permitted in {location} configuration"
+            )
+        else:
+            combined.append(f"- {formatted_msg} in field {formatted_loc!r}")
+
+    return "\n".join(combined)
