+++ tests/commands/test_pack.py
@@ -32,7 +32,8 @@ from charmcraft.bases import get_host_as_base
 from charmcraft.cmdbase import JSON_FORMAT
 from charmcraft.commands import pack
 from charmcraft.commands.pack import PackCommand, _get_charm_pack_args, _subprocess_pack_charms
-from charmcraft.config import load, BasesConfiguration
+from charmcraft.models.config import BasesConfiguration
+from charmcraft.config import load
 
 
 def get_namespace(
